requirements.txt
# Context Field Conditioning - Dependencies
# All local, no cloud services

# Core
numpy>=1.24.0
scipy>=1.10.0
requests>=2.28.0
pyyaml>=6.0

# Analysis
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Metrics
scikit-learn>=1.2.0      # for cosine similarity (GNS metric)
sentence-transformers>=2.2.0  # for embedding-based novelty scoring

# Statistics
statsmodels>=0.14.0      # for ANOVA, interaction analysis
pingouin>=0.5.0          # for effect sizes, Bonferroni

# Utilities
tqdm>=4.65.0             # progress bars for 1000 runs
jsonschema>=4.17.0       # validate honeycomb cells against schema
